+++ OhmEngine/include/OhmEngine/RendererVulkan/FrameBuffer.hpp
@@ -1,0 +1,20 @@
+#pragma once 
+
+namespace OHE
+{
+    class FrameBuffer
+    {
+        public:
+            FrameBuffer();
+            ~FrameBuffer();
+
+            FrameBuffer(const FrameBuffer&) = delete;
+            FrameBuffer& operator=(const FrameBuffer&) = delete;
+
+            void CreateFramebuffers();
+            void DestroyFramebuffers();
+
+        private: 
+            std::vector<VkFramebuffer> m_swapChainFramebuffers;
+    };
+}
