+++ OhmEngine/include/OhmEngine/RendererVulkan/PhysicalDevice.hpp
@@ -15,13 +15,6 @@ namespace OHE
         }
     };
 
-    struct SwapChainSupportDetails
-    {
-        VkSurfaceCapabilitiesKHR capabilities;
-        std::vector<VkSurfaceFormatKHR> formats;
-        std::vector<VkPresentModeKHR> presentModes;
-    };
-
     static std::vector<const char *> deviceExtensions = {
         VK_KHR_SWAPCHAIN_EXTENSION_NAME,
     };
@@ -32,51 +25,20 @@ namespace OHE
         PhysicalDevice();
         ~PhysicalDevice();
 
-        WindowSurface &GetWindowSurface() { return windowSurface; }
         void PickPhysicalDevice(VkInstance &instance);
         void CreateLogicalDevice(bool enableValidationLayers);
-        void CreateSwapChain();
-        void CreateImageViews();
         void DestroyLogicalDevice();
-        void DestroySwapChain();
-        void DestroyImageViews();
-        void CreateRenderPass();
-        void DestroyRenderPass();
-        void CreateFramebuffers();
-        void DestroyFramebuffers();
-        void CreateCommandPool();
-        void DestroyCommandPool();
-        void CreateCommandBuffer();
-        void CreateSyncObjects();
-        void CleanupSyncObjects();
 
-        void DrawFrame();
         void WaitIdle() { vkDeviceWaitIdle(device); };
-
-        void RecordCommandBuffer(VkCommandBuffer commandBuffer, uint32_t imageIndex);
 
         //Temporary
         VkDevice &GetLogicalDevice() { return device; }
-        VkRenderPass &GetRenderPass() { return renderPass; }
 
     private : 
         VkPhysicalDevice physicalDevice = VK_NULL_HANDLE;
         VkDevice device;
         VkQueue graphicsQueue;
         VkQueue presentQueue;
-        WindowSurface windowSurface;
-        VkSwapchainKHR swapChain;
-        std::vector<VkImage> swapChainImages;
-        VkFormat swapChainImageFormat;
-        VkExtent2D swapChainExtent;
-        std::vector<VkImageView> swapChainImageViews;
-        VkRenderPass renderPass;
-        std::vector<VkFramebuffer> swapChainFramebuffers;
-        VkCommandPool commandPool;
-        std::vector<VkCommandBuffer> commandBuffers;
-        std::vector<VkSemaphore> imageAvailableSemaphores;
-        std::vector<VkSemaphore> renderFinishedSemaphores;
-        std::vector<VkFence> inFlightFences;
 
         const int MAX_FRAMES_IN_FLIGHT = 2;
         uint32_t currentFrame = 0;
@@ -84,13 +46,5 @@ namespace OHE
         bool IsDeviceSuitable(VkPhysicalDevice device);
         bool CheckDeviceExtensionSupport(VkPhysicalDevice device);
         QueueFamilyIndices FindQueueFamilies(VkPhysicalDevice device);
-
-        SwapChainSupportDetails QuerySwapChainSupport(VkPhysicalDevice device);
-
-        VkSurfaceFormatKHR ChooseSwapSurfaceFormat(const std::vector<VkSurfaceFormatKHR> &availableFormats);
-        VkPresentModeKHR ChooseSwapPresentMode(const std::vector<VkPresentModeKHR> &availablePresentModes);
-        VkExtent2D ChooseSwapExtent(const VkSurfaceCapabilitiesKHR &capabilities);
-
-        VkShaderModule CreateShaderModule(const std::vector<char>& code);
     };
 } // namespace OHE
