+++ OhmEngine/include/OhmEngine/RendererVulkan/RenderPass.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+namespace OHE
+{
+    class RenderPass
+    {
+    public:
+        RenderPass(/* args */);
+        ~RenderPass();
+
+        void CreateRenderPass();
+        void DestroyRenderPass();
+
+    private:
+        VkRenderPass renderPass;
+    };
+}
