+++ OhmEngine/src/RendererVulkan/FrameBuffer.cpp
@@ -1,0 +1,44 @@
+#include <OhmEngine/RendererVulkan/FrameBuffer.hpp>
+
+namespace OHE
+{
+    FrameBuffer::FrameBuffer()
+    {
+    }
+
+    FrameBuffer::~FrameBuffer()
+    {
+    }
+
+    void FrameBuffer::CreateFramebuffers()
+    {
+        m_swapChainFramebuffers.resize(swapChainImageViews.size());
+        for (size_t i = 0; i < swapChainImageViews.size(); i++)
+        {
+            VkImageView attachments[] = {
+                swapChainImageViews[i]};
+
+            VkFramebufferCreateInfo framebufferInfo{};
+            framebufferInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
+            framebufferInfo.renderPass = renderPass;
+            framebufferInfo.attachmentCount = 1;
+            framebufferInfo.pAttachments = attachments;
+            framebufferInfo.width = swapChainExtent.width;
+            framebufferInfo.height = swapChainExtent.height;
+            framebufferInfo.layers = 1;
+
+            if (vkCreateFramebuffer(device, &framebufferInfo, nullptr, &m_swapChainFramebuffers[i]) != VK_SUCCESS)
+            {
+                throw std::runtime_error("failed to create framebuffer!");
+            }
+        }
+    }
+
+    void FrameBuffer::DestroyFramebuffers()
+    {
+        for (auto framebuffer : m_swapChainFramebuffers)
+        {
+            vkDestroyFramebuffer(device, framebuffer, nullptr);
+        }
+    }
+}
